parsing test commands now
A few logic scripts are failing because the size of the ELSE
clause appears larger than the end of the script.  I wonder
if I should just convert those to a plain GOTO and continue on,
in those cases?  I'm not yet sure what's going on there.  Maybe
it's a nested IF, and the "ELSE" is jumping across the outer
IF/THEN/ELSE boundary?  unstructured hijinx.