Refactoring Engine, preparing to load resource bytes.
Now the Game class is gone, and the Engine (renamed from EngineVersion)
takes over the Facade-like responsibility for putting an easy-to-use
face on all of these objects.

An Abstract Factory (ResourceLoadingFactory) helps the engine use the
right set of objects to accomplish this work, hopefully making it easy
for programs to use, e.g., wrapped debug-versions of the classes.
Though it's not planned, it should also make it feasible to mock out the
filesystem access for unit testing.