Created enough of a resource loader to dump raw bytes of sounds.
The next step is to define a Sound object which doesn't presuppose
what the client wants to do with the sound.  The idea is to basically
provide an interpreter for the sound bytes that will push data into
a client-provided builer, which will build whatever the client needs
from it.  We'll see how that works in practice...