Refactoring yesterday's work.
I realized that in almost all cases, only one instance of
most classes handled by the factory would ever be created.
As such, the factory setup was overkill, and I converted to
a series of createDefault() static methods on the interfaces
that the client can use during setup.

This still allows the client to wrap or subclass the defaults
if needed, so no flexibility is lost.