Fixed up orchestras/Added entire KQ3 sound catalog.
I've got a set of csound orchestras I'm really happy with now,
as the scores reflect the actual data in the AGI resources, rather
than being "pre-interpreted" in terms of amplitudes and frequencies.
This will make it much easier to enhance the sound output via the
orchestra files, and leave the scores alone.

I now have to 1) re-extract all the sound resources from all the games,
and 2) re-add whatever orchestration I want to the "favorites" I have.