Added PIC resource decoding.
I used the line-drawing algorithm from
http://agiwiki.sierrahelp.com/index.php?title=AGI_Specifications:_Chapter_7_-_Picture_Resources
... but it may not be 100% correct.  Almost every screen I look at is
right, but a couple seem to flood-fill too far.  So I will check out
problem cases and glance at the scummvm code to get it 100% right.

Also, while it's not dreadfully slow, it's slower than I expected.  So I
may profile it and see about speeding it up... or maybe I'll wait and do
that if I every actually go the extra step to make an AGI interpreter.