Added the AGI lzw decoder as two transducers.
This is a really nice decomposition of the algorithm, so I'm
pleased with what clojure could do here.  One transducer splits
the LZW stream into tokens, and the other one expands the tokens.
Hopefully the performance is acceptable.  My java version writes
directly into the output array and does a mutating reversal in
place.  The transducer version writes to a transient vector, and
then reduces over the `rseq` of it down the transducer chain.

Another performance enhancement was: since I know the usage is
single-threaded, I made the transducer state unsynced-mutable
by putting it into a deftype implementing IFn.