Added first ability to pull resources from VOL files.
This involves adding a "transparent" open file cache, so you
don't re-open the same files over and over again while pulling
resources.  I'm happy with the implementation, which should do
the job well enough in only a few lines of code.

I've structured the functions so that you'll always have low-level
access to the raw (decompressed) bytes of the resource if you just
call `(load-resource-bytes game (nth (:sounds game) 21))`, for
example.  So, if your application really doesn't want the default
formatting of the resources, you can still use the library to get
the raw bytes.