Added MIDI output for extracted songs.
This uses the javax.midi package, rather than writing raw midi
data like the CSharp version.  I may add cmdline args to select
instruments and panning for the channels, or maybe it's easy
enough to just open the MIDI file in a sound program and adjust
those settings to taste.