upgrading dependencies and renaming namespaces
Preparing for a refactor where the agires lib can go to maven
and not depend on java.desktop.  The main consequence will be
that it won't give back bufferedimages.  Instead, arrays of
bytes will suffice.  The receiver will know what to do with
them (whether the images are destined for swing or javafx or
libGDX they will need to be converted differently anyway).