Moved disassembler into agires library.
This moves the disassembler into the agires library, and makes
the disassembly available from any AgiLogicScript that you
load.

I considered changing the output from a print-to-PrintWriter
approad to returning more like an AST... but for now I'm holding
in a YAGNI pattern.  After all, for a makeshift debugger I can
parse the lines of the disassembly to identify good breakpoint
targets, etc.  We'll see when we get there!