Converted view handling to AgiView class.
Slowly, I'm doing away with the builders of yore, and returning
concrete data-holding classes to the client.  And, I'm still slowly
getting the agires library decoupled from java.desktop.  Only PIC
resources are left to convert.

Then, I'll go back and revisit the Sound Builder, and have it return
a data class of the song structure, and build MIDI and Csound off of
that instead.  I'm doing that just for consistency, since it will
be the only Builder left.