Almost finished moving away from the 'builder' system.
Now all resource types return an abstract representation of the
resource, which the client can manipulate however they want.

The one feature I still need to re-add is the ability to poll the
drawing of PIC resources, so that time-lapse movies of the drawings
can be made.

Then, I'll need to add javadoc everywhere it is missing in the library,
and add it to Maven/Sonatype.